+++ src/lazyPages.js
@@ -18,6 +18,7 @@ export const TodosList = lazy(() => import('@/pages/Todos/TodosList'));
 export const FeedbackList = lazy(() => import('@/pages/Feedback/FeedbackList'));
 export const FeedbackDetail = lazy(() => import('@/pages/Feedback/FeedbackDetail'));
 export const Communication = lazy(() => import('@/pages/Communication/Communication'));
+export const AppAccess = lazy(() => import('@/pages/AppAccess'));
 export const Settings = lazy(() => import('@/pages/Settings/Settings'));
 export const VOG = lazy(() => import('@/pages/VOG/VOG'));
 export const Contributie = lazy(() => import('@/pages/Contributie/Contributie'));